Add support for `wasmtime run --argv0 NAME ...`
This commit adds support for a new `--argv0` flag to the `wasmtime run`
CLI (and only `wasmtime run`, not other subcommands). This flag is used
to override the inferred value of the first argument which is sometimes
used to dispatch on various pieces of functionality. This enables
avoiding the need to copy around wasm files.

Closes #8955