Fix testing some crate/feature combinations in this workspace (#9002)
* Fix test of wasmtime-fuzzing with `fuzz-pcc` enabled

Allow pcc errors when generating fuzz modules since it's currently gated
as the errors haven't all been weeded out.

* Gate wmemcheck instrumentation more in `wasmtime-cranelift`

This fixes an issue where when the `wmemcheck` feature was enabled it
would unconditionally enable the `wmemcheck`-related instrumentation
around `malloc` and `free`. The fix is to additionally check the runtime
`wmemcheck` setting as well.

Closes #8997

* Forward the `wmemcheck` feature from `wasmtime-cranelift` to `wasmtime-environ`

Not necessarily strictly required but it should help prevent issues
such as #8997 in the future.