CI: Test Pulley on 32-bit and `no_std` targets (#9013)
* CI: Check that Pulley builds for `no_std` targets

* CI: Test Pulley on 32-bit targets

This adds CI testing for Pulley on the following 32-bit targets:

* `i686-unknown-linux-gnu`
* `armv7-unknown-linux-gnueabihf`

The previous commit added checks that Pulley builds for `no_std` targets.

Our existing test sharding means that it is already tested on a big-endian
platform (`s390x-unknown-linux-gnu`) as well as on different OSes (Linux,
Windows, and macOS).

So altogether we are now testing a pretty wide range of portability dimensions
for Pulley in CI:

* 32- vs. 64-bit architectures
* `std` vs. `no_std`
* Little- vs. big-endian
* Linux vs. Windows vs. macOS

I think our CI coverage for Pulley is in a pretty good place now!

Co-Authored-By: Alex Crichton <alex@alexcrichton.com>

---------

Co-authored-by: Alex Crichton <alex@alexcrichton.com>