Remove `V128Abi` and shrink the size of `Val` from 32 to 24 bytes (#9023)
This removes `V128Abi`, which is no longer needed since we removed the native
calling convention from Wasmtime. For the new internal representation of `V128`,
we use an unaligned byte array. This lowers the alignment requirements for
`V128` to 1 rather than 16, which means that `Val` no longer requires padding
after its discriminant before the storage for its `V128` variant, which in turn
allows its size to shrink from 32 to 24 bytes.

Co-authored-by: Trevor Elliott <telliott@fastly.com>