Add a more first-class error for Wasmtime on 32-bit
This commit adds a more first-class error in the build script for the
`wasmtime` crate to indicate that 32-bit platforms are not supported.
While there is a `compile_error!` with somewhat similar information it
is not too helpful because the compiler keeps compiling things after the
`compile_error!` is reached meaning that the absence of a `sys` module
cascades to many other compile errors as well. This presents a
more readable error with links to github issues in a way that halts the
build and should prevent unsupported crates from being built.