s390x: Support tail calls (#9052)
This adds support for a newly defined tail-call ABI for s390x
as well as supporting tail calls themselves.

We now use the tail-call ABI by default for Wasmtime, and
enable tail calls by default.

This also allows getting rid of a number of special case
and test exclusions for s390x.

Fixes: https://github.com/bytecodealliance/wasmtime/issues/6530