cranelift-frontend: Make `assert_eq_output!` usable in the whole crate (#9070)
This moves the macro to `lib.rs` so that it can be used in all the tests in the
whole crate. It also makes it a little more usable by trimming whitespace and
printing both the expected and actual values, in addition to the diff.

A follow-up commit will migrate various uses of `assert_eq!` over to it.

Co-authored-by: Trevor Elliot <telliott@fastly.com>