Update to WASI WITs 0.2.1 (#9063)
No major changes happen in this release so it's just bumping version
numbers. The goal here is to turn the crank and ensure that nothing else
breaks as part of this release process. There's a surprising number of
versions to update in-repo which we may want to make easier in the
future but for now this just manually updates.

It is planned that this does not actually break anyone in practice.
Older runtimes should support newer versions and Wasmtime after this PR
should continue to support older binaries as well.

Put another way this should not break things and if it does this will
get reverted and fixed. We've done our best to make sure this won't
break things, so let's find out in the real world now.