Update handling of wasm features and `*.wast` testing
This commit is an attempt at making it easier to manage the set of
variables we have in play with `*.wast` testing and wasm features.
The Cranelift and Winch backends support different sets of wasm features
at this time and historically Cranelift has also had
architecture-specific support for wasm features. This is hoped to help
simplify management of all of this in the future in addition to making
`*.wast` testing more robust. Notable changes here are:

* A `Config` no longer tracks a `WasmFeatures` but instead only tracks
  explicitly disabled/enabled features. The final `WasmFeatures` is then
  computed given the result of compiler configuration.
* Default-enabled features now start from nothing instead of
  `wasmparser`'s defaults to avoid future breakage there.
* Each compiler configuration and/or backend now has a listed set of
  "this feature may panic" wasm features. These are all disabled by
  default and if explicitly enabled a `Config::validate` error is
  returned.
* All `*.wast` tests are now run through both Cranelift and Winch. Tests
  are now asserted to either fail or pass depending on configuration and
  the whole test will fail if the result doesn't match the expectation.
* Many `gc` proposal tests which are now passing are flagged as passing
  now. Failing `winch` tests are now explicitly listed instead of using
  a litany of patterns.