Do not deallocate disabled GC heaps via the instance allocator (#9180)
* Do not deallocate disabled GC heaps in the instance allocator

When the reference types proposal is disabled, we don't allocate a GC heap from
the instance allocator, we instead create a `DisabledGcHeap`. Rather than
returning this to the instance allocator when we're done with it, which won't
even know what to do with it since the instance allocator didn't create the GC
heap, we simply rely on the `DisabledGcHeap`'s `Drop` implementation.

* check gc types instead of reference types