aarch64: Add special-case for widening multiplication
This commit adds a special case to the lowering of 128-bit
multiplication on the aarch64 backend along the same lines as was done
in #9136 for the x64 backend. Notably zero and sign-extended values
which are multiplied to produce a 128-bit result can skip some of the
arithmetic of the fully general 128-bit lowering.