Fix Intel Mac CI builds. (#118)
* Fix Intel Mac CI builds.

We build both x86-64 and aarch64 ("Intel Mac" and "Apple Silicon Mac")
binaries for wizer in CI on the `macos-latest` CI runner. Historically,
this runner was an x86-64 machine, so while we could do a direct compile
for x86-64 binaries, we added a target override for `aarch64-darwin` for
the aarch64 builds to force a cross-compile.

When GitHub switched macOS CI runners to aarch64 (ARM64) machines
somewhat recently, the `macos-latest` runner image began producing
aarch64 binaries by default, and the target override for
cross-compilation became meaningless (redundant). As a result, *both* of
our x86-64 and aarch64 macOS binary artifacts contained aarch64
binaries. This is a problem for anyone running an Intel Mac.

This PR switches the CI config to specify a cross-compilation of x86-64
binaries, so each target builds its proper architecture.

* Use actions/upload-artifact v4

* Use v4 of actions/download-artifact too