vtune: smoke test profiling in CI (#9185)
* vtune: check that VTune profiling works

This adds a smoke test for checking that the VTune functionality still
works.
- if `vtune` is not on the system path, skip the test
- if it is, run Wasmtime with VTune profiling enabled
- check that it captured the CPU time and didn't fail.

This test doesn't rigorously check all the ins and outs of the VTune
functionality but is better than before.

* ci: install VTune

In order to actually run the VTune CLI test, we need `vtune` on the
system. The `install-vtune-action` I created does this ([script]).

[script]: https://github.com/abrown/install-vtune-action/blob/main/install.sh

* ci: further restrict workflow runs

Running the VTune installation and CLI test on any `ubuntu` runner meant
that even test runs for `riscv64` (emulated on QEMU) would attempt this.
Instead, restrict this test only to `linux-x64` runs.

prtest:full

* review: add `cfg(target_arch)`