Remove CLIF `umulhi` and `smulhi` instructions
Instead of supporting these instructions natively, which for example
most source languages don't have direct support for (nor wasm),
pattern-match these instructions during lowering. This commit replaces
backend-specific lowerings of `umulhi` and `smulhi` with patterns in
each backend in respective places. This helps detect when this happens
from WebAssembly for example and helps simplify producers as well.
The intent of this commit is that when desired the same instructions are
emitted as before.

The downside of this commit is that CLIF is losing the ability to
implement `umulhi` and `smulhi` on vector types. This was only
implemented in the riscv64 and s390x backends and was not used by
WebAssembly for example.