Make host trap handlers optional
This commit introduces a new configuration option
`Config::host_trap_handlers` which enables disabling the reliance on
host trap handlers (e.g. signal handlers on Unix) at runtime. This
is intended to help increase portability in cases where signal handlers
are otherwise difficult. This is achieved by plumbing the translation
environment to more locations which now conditionally lowers to calls to
a function to raise a trap instead of a trap instruction.

The caveats of this support are:

* This is not yet implemented for Winch
* This requires disabling spectre mitigations
* This does not yet support shared memories since it forces dynamic
  memories to be used.

These points are all possible to address but it might be best to see how
this feature evolves over time. I'll also note that this is not an
optimized implementation and likely has a fair bit of overhead. For
example the solution in #6926 of a more optimized stub to jump-and-trap
is not implemented yet.

Closes #6926