Revert back to 1.78 as workspace MSRV
Last night's OSS-Fuzz build failed because the compiler used there
hasn't been updated in quite some time. It's a nightly of 1.78 from the
beginning of this year which means the bump to 1.79 recently for this
workspace meant that Wasmtime stopped building there.

This commit reverts ac607ed2abdf0d51aedca302a78aa27d03451d2c and
1ad3da85ef8d161071b57aeed86effdd70cbc1ce and then updates our github
action to say that we're 3 versions behind the latest Rust right now
with some commentary as to why.