Remove some job dependencies in CI
All of the `cargo check`-style jobs are always run on both PRs and merge
so there's no need to gate on the `determine` step. That's only needed
when the output of `determine` conditionally decides to run a job or
not, but these are always unconditionally run. This will help a tiny bit
with CI time where the `determine` job doesn't need to run, but
otherwise cuts down a bit on the amount of configuration.