Don't limit core instances/memories/tables by default
This commit updates the defaults of `PoolingAllocationConfig` to allow
unlimited core instances/memories/tables per component instead of the
previous default of 20. The limits here are somewhat restrictive and
more tailored for high-scale use cases which aren't always the best
defaults for a general purpose tool like the `wasmtime` CLI. The CLI for
example uses the pooling allocator by default with `serve` and this
helps accept a wider breadth of components by default instead of
artificially limiting the shape of input components.